Ensure disconnected listeners are gc'able
Ray pointed out that since SnapshotArrayList was holding onto a reference in everything in the
snapshot between dispatches, disconnected listeners would be held in memory till the next dispatch.
Switch to CopyOnWriteArrayList to prevent that while keeping the same ability to modify the list
while iterating over it.

CopyOnWriteArrayList doesn't work in Collections.sort, so instead insert new listeners in priority
order with a binary search. This has the benefit of not resorting the already sorted list on every
addition.